+++ CHANGELOG.md
@@ -10,6 +10,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Fixed
 
+
+## [1.13.0] - 2021-11-30
+### Added
+- Add support for creating/importing projects with multi-member districts [#1060](https://github.com/PublicMapping/districtbuilder/pull/1060) 
+- Add Map Configuration  modal [#1065](https://github.com/PublicMapping/districtbuilder/pull/1065)
+### Changed
+- Update Competitiveness metric on Evaluate [#1068](https://github.com/PublicMapping/districtbuilder/pull/1068)
+### Fixed
+- Improve PlanScore integration and display toast on errors [#1062](https://github.com/PublicMapping/districtbuilder/pull/1062)
+- Fix duplicate regions in Community Maps dropdown [#1070](https://github.com/PublicMapping/districtbuilder/pull/1070)
+
 ## [1.12.1] - 2021-10-29
 ### Fixed
 - Fix import for regions that have a geounit with 1 identical geometry nested under it [#1053](https://github.com/PublicMapping/districtbuilder/pull/1053)
@@ -361,7 +372,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[unreleased]: https://github.com/publicmapping/districtbuilder/compare/1.12.1...HEAD
+[unreleased]: https://github.com/publicmapping/districtbuilder/compare/1.13.0...HEAD
+[1.13.0]: https://github.com/publicmapping/districtbuilder/compare/1.12.1...1.13.0
 [1.12.1]: https://github.com/publicmapping/districtbuilder/compare/1.12.0...1.12.1
 [1.12.0]: https://github.com/publicmapping/districtbuilder/compare/1.11.0...1.12.0
 [1.11.0]: https://github.com/publicmapping/districtbuilder/compare/1.10.1...1.11.0